Adds byoip support to the smartcontract/sdk/cli
Smartcontract user account updated with user_type that supports the
enumerations we want to support (ibrl, ibrl with allocated ip, edge
filtering and multicast). Activator modified to not allocate an ip for
ibrl. CLI gets --allocate-ip option and a dz-mode (ibrl, edgefiltering,
multicast) as well as sends the user_type to doublezerod.