Activator test bug fix
The activator tests were not structured correctly and passing when they
shouldn't have been (due to a malformed test, not an activator bug). The
mockall crate has a pecularity whereby when expect_foo is called to
expect a function call to foo, it matches any expectation in the list
that meets the criteria. In this case, a call to
expect_execute_transaction with out specific params on it caused
the tests to pass when the wrong expectations were occurring.
Eliminating this generic expectation fixed the tests so that they broke,
and then the params of the tests were adjusted to exercise the correct
behavior in the Activator.