DZ #549: allow sla-program integrations (#550)
Resolves #549.

# What changed?

Added "no-entrypoint" feature. This change allows the
doublezero-sla-program package to be added to another program's package
(which has a defined entrypoint). Without it, the BPF compiler cannot
build the integration's program due to conflicting entrypoints.

For reference:
https://docs.rs/solana-program-entrypoint/latest/solana_program_entrypoint/macro.entrypoint.html#cargo-features.

There is no evidence that the change makes a difference because there
are no example integrations in this repo. But I tested this myself in
another repo and I can successfully build my program.

This PR also attempts to follow
https://github.com/malbeclabs/doublezero/pull/477.