Spawn worker on start instead of each loop (#554)
## Summary of Changes

This PR moves the client spawning out of the loop so that a single
client can be reused. Previously a new client would be started on each
iteration of the loop because `nil` was passed in for a new clientConn.
Moving the client start into the main loop prevents that from happening
and relies on grpc's [retry
logic](https://github.com/grpc/proposal/blob/master/A6-client-retries.m)
to handle connection failures.

## Testing Verification 

* Tests have been updated with the new function signature and the
removal of the `error` that is no longer needed
* Tests that test failure conditions have been removed on the client
because it's no longer possible to return an error from a new client
instantiation (the error would be triggered in the body of the `main`
function