Pull Request Template (#477)
## Summary of Changes
This PR proposes a standard PR template for our PRs. The intent is to
provide a standard, consistent way to describe changes and evidence for
those changes. Documentation should live as close to the code as
possible and then can be referenced in the future. Changes should be
written in complete sentences or bulleted points that follow standard
grammatical rules. While this might seem onerous, I've found that the
clearer, in writing, one expresses oneself, the better one understands a
thing and has the benefit of making it easier to explain to someone.
Since we will be building in public and working closely with core
contributors like Anza and Firedancer, this is all the more important.

These changes could perhaps be part of an RFC of how we use git at DZ,
but these changes in and of themselves don't warrant an RFC as they are.

### Commit Message
I'd also like to propose a commit message _to main_ of the format `DZ
((#\d+\s?)+)-\s(.+)`. Basically, `DZ #123 (one or more issues) - some
meaningful, succinct overview of change`. Unfortunately github issues
are id'ed with `#` which causes issues when rebasing as it's read as a
comment so we need a prefix - a space, a !, or DZ or something else.
Having the issue in the commit message helps keep a low cost
traceability throughout the repo and makes the commit history more
valuable and searchable (ie why did we do this). We can enforce this
with git hooks to ensure compliance.

### Signing Commits
Additionally, I'd like to propose that we all sign our commits so that
the commits are `Verified`. It's trivial to add signing with an [SSH
key](https://docs.github.com/en/authentication/managing-commit-signature-verification/about-commit-signature-verification#ssh-commit-signature-verification)
and helps prove that the author is who we think the author is.