release: deploy dailys to devnet (#671)
## Summary of Changes

This PR adds a deployment job to the already present daily release
github action, which will deploy the previously cut daily release to
devnet. The smart contract deployment will be handled as a follow up PR.
Also included is a change to the version format for these packages as
the previous format only included the git hash in the package name which
created issues with apt/yum determining the latest release since they
compare lexicographically given two releases with the same version
number. The new package format includes a timestamp of when the release
was built as well as the git hash.

## Testing Verification

This ran successfully across all components (minus smart contract deploy
which isn't included):

https://github.com/malbeclabs/doublezero/actions/runs/16013528138