Change Management RFC (#665)
## Summary of Changes
This PR introduces the change management RFC which provides a foundation
for standardized and regular releases, versioning, and automated testing
in remote environments. Ultimately, this RFC is the lodestar for the
`Weekly release to testnet` milestone in the [mainnet-beta requirements
notion
page](https://www.notion.so/malbeclabs/214fef22bebe80be9c69c92236806cb3?v=214fef22bebe800f8ad6000c37a50e34).

Note that the rfc ID will be updated when it's ready to merge. 
## Testing Verification
N/A