agent: expose optional prom metrics (#746)
## Summary of Changes
This PR adds optional prometheus metrics to the device agent with can be
exposed via a `-metrics-enable` flag. The following metrics are exposed:

- `doublezero_agent_build_info` - Build information of the agent
(version, commit, date)
- `doublezero_agent_bgp_neighbors_errors_total` - Number of errors
encountered while retrieving BGP neighbors from the device
- `doublezero_agent_get_config_errors_total` - Number of errors
encountered while getting config from the controller
- `doublezero_agent_apply_config_errors_total` - Number of errors
encountered while applying config to the device

The main use case for this is to derive the health of a release to
devnet as well as tracking the current build running on each device. But
this may also be valuable to contributors who would like to integrate
the agent into their existing monitoring stack, assuming they use
prometheus.

Also included is an adjustment to the version added to the agent via
LDFLAGs as we need this to match the version set by goreleaser in the
package so when we build a pre-release of the next version increment,
this is reflected in the build (which is then exposed via prometheus).
In many cases, the version field was empty if the current commit is not
part of a git tag, which is the case when we cut a pre-release.

## Testing Verification

Started the agent outside of a device and watched errors increment. The
build info is also correct:
```
ubuntu@chi-dn-bm7:~$ curl -s localhost:8080/metrics | grep doublezero
# HELP doublezero_agent_apply_config_errors_total Number of errors encountered while applying config to the device
# TYPE doublezero_agent_apply_config_errors_total counter
doublezero_agent_apply_config_errors_total 0
# HELP doublezero_agent_bgp_neighbors_errors_total Number of errors encountered while retrieving BGP neighbors from the device
# TYPE doublezero_agent_bgp_neighbors_errors_total counter
doublezero_agent_bgp_neighbors_errors_total 4
# HELP doublezero_agent_build_info Build information of the agent
# TYPE doublezero_agent_build_info gauge
doublezero_agent_build_info{commit="eac3a78a",date="2025-07-10T02:28:12Z",version="0.2.3~git20250710022812.eac3a78a"} 1
# HELP doublezero_agent_get_config_errors_total Number of errors encountered while getting config from the controller
# TYPE doublezero_agent_get_config_errors_total counter
doublezero_agent_get_config_errors_total 4
```