smartcontract: fix account creation (#748)
## Summary of Changes

### What changed?
- Add `doublezero-program-common` crate
- This crate can be used to share common utilities between
Serviceability and Telemetry programs
- Add `try_create_account`, which follows a prescribed method to prevent
DOS when creating accounts
- Replace all instances of invoking the create-account instruction with
`try_create_account`

### Why is this change necessary?

Without this change, an attacker can send some number of DZ Ledger SOL
to a PDA, which would result in an error resembling:
```sh
[2025-07-10T16:31:12Z DEBUG solana_runtime::message_processor::stable_log] Create Account: account Address { address: 5X57hSVpkLA446YBmVSM2ZwWGh4h7ST1YEVMNuSxvdcY, base: None } already in use
[2025-07-10T16:31:12Z DEBUG solana_runtime::message_processor::stable_log] Program 11111111111111111111111111111111 failed: custom program error: 0x0
[2025-07-10T16:31:12Z DEBUG solana_runtime::message_processor::stable_log] Program 1111111QLbz7JHiBTspS962RLKV8GndWFwiEaqKM consumed 43966 of 1000000 compute units
[2025-07-10T16:31:12Z DEBUG solana_runtime::message_processor::stable_log] Program 1111111QLbz7JHiBTspS962RLKV8GndWFwiEaqKM failed: custom program error: 0x0
```

PDA keys can be known prior to its account being created. So with the
Serviceability program for example, because all PDAs use an index
derived from a global index, an attacker can send lamports to the
program account for the next index. Once this happens, no more accounts
can be created on the Serviceability program.

## Testing Verification
- Added one test on the Telemetry program to test the fix. Other tests
resembling this test can be added to ensure any instructions that create
accounts use the `try_create_account` method.