activator: add shutdown signal listening (#813)
## Summary of Changes
* Wrap the activator's rust processes in a biased tokio select that
awaits a shutdown synchronization primitive, deferring to the shutdown
future in the scheduler in the event an exit signal is sent from the
server's host OS and then gracefully cancels the activator sync
subscriber handler and metrics submitter processes.
* For a long running server process, we should listen for and handle
exit signals propagated from the host OS to gracefully shut down the
application processes and prevent it from hanging and/or requiring the
host systemd from having to force-kill the process on timeout when it's
attempting to shut down for any reason.

## Testing Verification
* I'm unclear if there's a proper way to test this change in the e2e
suite but in theory we can ensure the exit signals from the test runner
are propagated to the activator server process by wrapping the test
dockerfile's invocation of the entrypoint.sh script in a call to `tini`

addresses https://github.com/malbeclabs/doublezero/issues/831