Get inflation rewards for some number of validators (#845)
## Summary of Changes

This PR gets `inflation_rewards` for `n` number of validators.
Curiously, to get inflation rewards, you first need to get the vote
pubkey for the validator so you have to `get_vote_accounts_with_config`
to get all the vote keys and the associated validator pubkeys so you can
then use the `vote_pubkey` to get the rewards for that validator (which
is as cumbersome as it sounds).

The `get_rewards` function will ultimately return the sum of rewards
along with their sources for each validator.

## Testing Verification
* Simple unit test to validate the response has the expected validator
pubkey and reward amount