Calculate revenue for validator between two timestamps (#885)
## Summary of Changes

This PR adds the ability to generate total revenue between two
timestamps. It takes in `start` and `end` ios8601 timestamps and then
converts each timestamp to a solana epoch. Then it gets the range of
epochs from the start to the end epoch. Then it loops over the epochs
and calculates the rewards for each epoch with the `get_rewards`
function.


## Testing Verification
* Unit test was added to validate that the expected epochs have been
returned