Mock JITO api (#920)
## Summary of Changes
This PR adds a mock JITO api to ensure the tests are speedy and don't
depend on epochs that change and older data can become unavailable. It
uses `mockall` to do the mocking by creating a trait. Then `mockall`
macros fill out the rest. Unfortunately, it doesn't seem to be easy to
get a mock solana client so it might be more worthwhile to simply use
reqwest and then pass in the mockable client. This would also clean up
the API to use a single client instead of two.


## Testing Verification
* unit test still validates jito rewards except it uses the mock API now
and still passes