get_total_rewards and get_rewards_between_two_timestamps (#950)
## Summary of Changes
This PR adds a wrapper around the solana client (thanks @karl-dz!) and
simplifies the interface. It adds in an `RpcBlockConfig` arg to the
`get_block_rewards` function to provide more control over what's
returned from the caller. Likewise it updates `get_inflation_rewards` to
use a mock with the same `ValidatorRewards` interface.

This closes https://github.com/malbeclabs/doublezero/pull/924. 

https://github.com/malbeclabs/doublezero/pull/936 will complete the
first version of the API by adding in the full client for the 2 public
functions and then adding in mocks for those two functions.

This also fixes a few deprecation warnings.

## Testing Verification
* Unit test now mocks the remote API calls