release: automated qa testing (#1018)
## Summary of Changes
This PR adds a QA agent gRPC service to run on devnet or testnet nodes
in order to perform automated testing on baremetal hardware and will
ideally replace the manual tests ran by humans in devnet before
greenlighting a release to testnet as part of the daily release rollout.
`qa_test.go` provides example usage, which is to write a standard go
test using each hosts RPC client to orchestrate the test. The initial
RPCs contain implementations to connect to doublezero using ibrl mode,
the ability to send a set of pings, get the host's doublezero tunnel
status, and disconnect from doublezero. A multicast test will be added
after this PR.

## Testing Verification

Since you can't run a new github workflow from a branch, this is the
result of the qa test running as part of the daily release workflow,
with the release steps commented out:


https://github.com/malbeclabs/doublezero/actions/runs/16665787243/job/47171675639