release: fix dependency tree and deflake (#1032)
## Summary of Changes
This PR fixes the dependency tree for the daily deploy to make sure all
components are released prior to running QA tests. This also fixes
flakiness when running inside the runner as the 10s context timeout for
ping results can be tripped due to the RPC call initiation/1 second
delay between ping packets and the RPC reply.

## Testing Verification

https://github.com/malbeclabs/doublezero/actions/runs/16694719770