release: multicast qa test (#1036)
## Summary of Changes
This PR adds a multicast pub/sub test to automated QA in devnet. The
test chooses a random publisher with the remaining hosts as subscribers,
creates a multicast group onchain, publishes to the group and verifies
we received non-zero packet counts on all subscribers. With this test,
we should now have all our manual QA testing prior to a testnet release
automated and will be performed immediately after our daily devnet
rollout.

## Testing Verification


https://github.com/malbeclabs/doublezero/actions/runs/16713915195/job/47303701339