passport: access flow instructions and basic testing (malbeclabs/doublezero-solana#16)
Add logic for DoubleZero Passport program to request, grant and deny
access.

The DoubleZero Ledger Sentinel is a special authority that grants or
denies access to the DoubleZero Ledger network. It will either collect a
small fee for granting access or collect all of the lamports stored in
the request access account.

The user must provide a legitimate signature produced by a Solana
validator keypair. There is no verification of this signature or
authenticity of this validator ID onchain. The offchain sentinel will
perform the signature verification and determine whether to grant or
deny based on its success.

To disincentivize spam, a SOL deposit is required to make a request.
This amount will likely be configured to be much larger than rent
exemption lamports required to keep the access request account open.

Closes https://github.com/malbeclabs/doublezero/issues/1063.