controller: required network env flags (#1086)
## Summary of Changes
- Update controller to require the network env flags, rather than
defaulting to testnet, which is risky and error-prone on a component
like this that configures all the devices in the network based on just
read-only config to the ledger. Consider the scenario where the systemd
override is removed or doesn't exist in mainnet for whatever accidental
reason, and now all of the mainnet devices are configured for testnet
within seconds.
- Add support to the controller for the `--env` flag
- Remove controller .service file from packaging, just install the
binary, and let our ansible playbook manage everything else
- Related to https://github.com/malbeclabs/doublezero/issues/978

## Testing Verification
* Show evidence of testing the change