revenue-distribution: add verify distribution merkle root (malbeclabs/doublezero-solana#19)
Added an instruction to verify merkle roots posted to a distribution.
The purpose is to simulate a transaction calling this instruction for
each leaf to make sure the posted root is correct.

This PR only implements checking the Solana validator payments root. The
payments accountant can verify each leaf before it considers finalizing
these payments. Verifying reward shares will come in a subsequent PR.

This change also:
- Fixed a bug in sending transactions with banks client. Blockhash was
randomly the same so we would get "AlreadyProcessed" error.
- Small refactor of create account recipe.

Closes https://github.com/malbeclabs/doublezero/issues/1078.