add in env (testnet, devnet) as a param to cilent (#1164)
## Summary of Changes
This PR adds in an `env` param which allows a user to set `testnet` or
`devnet` instead of entering an `url`, `ws` and `program_id`. Once we
have mainnet-beta up, we'll add in a `mainnet` config.

There's some cleanup that can be done in the config but not necessary
for mainnet-beta. Tracked here. #1168
## Testing Verification
* e2e tests pass