feat(controller) #1162 add flag controlling whether interfaces and peers are rendered (#1169)
## Summary of Changes
Today when a device does not have both vpnv4 and ipv4 loopbacks, the
controller will ignore the device. This means:
- The controller will not render config for the device. User tunnels
can't be connected or disconnected.
- The controller will not configure the device as an ibgp/msdp peer on
other devices

In order to provide a smoother upgrade path in testnet, this PR adds a
flag that disables rendering of interfaces and loopbacks. With the flag
enabled, the controller will continue to render config, but it will not
include any device interfaces or ibgp/msdp peers in the rendered
configuration.

The flag is disabled by default.

Once the testnet upgrade is done and all necessary device interface
records have been created, and link records have been updated, the flag
can be enabled, and device interface and ibgp/msdp peer rendering will
commence.

We can also remove the flag to simplify the code when it's no longer
needed.

## Testing Verification
* Unit tests and manual testing.
* No e2e test

---------

Co-authored-by: Steven Normore <snormore@gmail.com>