revenue-distribution: add migrate instruction (malbeclabs/doublezero-solana#23)
Add migrate-program-accounts instruction.

Follow the commits to see the process of upgrading Solana devnet to fix
the program config account. Successful migration required two upgrades
(see linked issue for transaction hashes).

The migrate-program-accounts instruction is meant to be called by the
upgrade authority. With its current state, it is safe to execute this
multiple times (this instruction processor simply sets the migrated bit
to false).

This PR also adds a Solana devnet upgrade test, which deploys an upgrade
and calls the migrate instruction.

Closes https://github.com/malbeclabs/doublezero/issues/1195.