qa: publish qa agent package to testnet (#1232)
## Summary of Changes
This PR allows for the publishing of the QA agent package to our testnet
apt repo so the testnet/mainnet QA hosts are able to use only tagged
versions instead of having to add the devnet repo and use daily builds.

This also normalizes dots to dashes to match other multi-word services.

## Testing Verification

CI should pass.