revenue-distribution: add space for payments and claims (malbeclabs/doublezero-solana#37)
Before we start initializing distributions on Solana devnet and
finalizing merkle roots, we need to make sure that the distribution
account has space to account for Solana validator payments and reward
claims.

In subsequent PRs that handle these payments and rewards, a bit
indicating the index of a given Solana validator or contributor will be
flipped from 0 to 1 to indicate that a payment or reward has been
processed. Setting these bits prevent replays of these payments and
rewards.

This PR also adds more storage gap to the distribution in case we need
to add more fields to the distribution (like reward calculation
parameters).

Closes https://github.com/malbeclabs/doublezero/issues/1282.