controller: link metrics (#1468)
## Summary of Changes
This PR adds the ability for the controller to generate link metrics
based on delay measurements listed in onchain link data. Since the
onchain delay values are stored in nanoseconds, and the maximum metric
value we can configure is 2^24-2, delay values are converted to
microseconds and used directly as the metric value.

This also changes the behavior of rendering is-is configuration for an
interface, where an interface must be in an activated link and have a
valid metric. Previously, we were rendering is-is configuration for any
physical interface regardless of whether it was participating in a link.

## Testing Verification

The following is the diff of configuration of fra-dz001 in testnet
between a config generated from the current testnet controller and a
build from this PR:
```
ubuntu@nyc-tn-bm2:~$ doublezero link list | grep fra-dz001
 9kHVMSBtRs75mpchpDviP2dihQ32nTqoEPb73shn4DU5 | fra-dz001:fra-dz-001-x    | jump_       | fra-dz001    | Switch1/11/1      | fra-dz-001-x | Switch1/5/1       | DZX       | 10Gbps    | 2048 | 1.00ms   | 1.01ms    | 8         | 172.16.0.20/31 | activated | 66yfemxTAjCL686R4FFpGugx1myQ7X6m274MzWB82xBy
 AWHkNcwF7PSNCJq4vgypxuAYfyeyHTWMySGefSN7VpFn | lon-dz001:fra-dz001       | jump_       | lon-dz001    | Switch1/1/1.1004  | fra-dz001    | Switch1/1/1.1004  | WAN       | 10Gbps    | 9000 | 11.09ms  | 1.24ms    | 2         | 172.16.0.4/31  | activated | 66yfemxTAjCL686R4FFpGugx1myQ7X6m274MzWB82xBy
```
```
➜  controller git:(ss/isis_metrics) ✗ diff -U8 fra-dz001.old fra-dz001.new
--- fra-dz001.old       2025-08-28 23:57:55
+++ fra-dz001.new       2025-08-28 23:58:15
@@ -80,27 +80,29 @@
 interface Switch1/1/1.1004
    mtu 2048
    encapsulation dot1q vlan 1004
    ip address 172.16.0.5/31
    pim ipv4 sparse-mode
    isis enable 1
    isis circuit-type level-2
    isis hello-interval 1
+   isis metric 11090
    isis hello padding
    isis network point-to-point
 !
 interface Switch1/11/1
    mtu 2048
    no switchport
    ip address 172.16.0.20/31
    pim ipv4 sparse-mode
    isis enable 1
    isis circuit-type level-2
    isis hello-interval 1
+   isis metric 1000
    isis hello padding
    isis network point-to-point
 !
 interface Loopback1000
    description RP Address
    ip address 10.0.0.0/32
 !
 mpls ip
  ```